Logarithmic visualization scale and dot blur in transistion map

+++ dvs_calibration/include/dvs_calibration/transition_map.h
@@ -53,7 +53,7 @@ private:
 
   uint64_t last_off_map_[sensor_width][sensor_height];
   uint64_t last_on_map_[sensor_width][sensor_height];
-  uint64_t transition_sum_map_[sensor_width][sensor_height];
+  double transition_sum_map_[sensor_width][sensor_height];
 
   CalibrationParameters params_;
